Use correct permission 1777 with sticky bit for /dev/shm
Permission on /dev/shm is always "drwxrwxrwt root root"
but this was changed to 0700 when mounting a memory
volume over /dev/shm, loosing the sticky bit and the
ability for an app running as a user to use shm_open.
So this patch adds a special case to give /dev/shm
the correct permissions 1777 that relies on the sticky
bit.