Removed empty environment variable constraint
What is the reasoning behind this check? I found out about this, because me and my colleague were debugging a pipeline for hours to find out, that one of my environment variables was actually missing. The value of the variable was `""`, an empty string literal, which is a valid value for environment variables. From the input of this function it looks to me like a string type check is not needed as a replacement. 
- @Alexander-Zierhut